# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CampaignDetail": {
        "properties": {
          "actor_names": {
            "items": {
              "type": "string"
            },
            "title": "Actor Names",
            "type": "array"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/CampaignEventRead"
            },
            "title": "Events",
            "type": "array"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "stages": {
            "items": {
              "type": "string"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "enum": [
              "active",
              "inactive",
              "closed"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "first_seen",
          "last_seen",
          "status",
          "event_count",
          "actor_names",
          "stages",
          "version",
          "created_at",
          "updated_at",
          "events"
        ],
        "title": "CampaignDetail",
        "type": "object"
      },
      "CampaignEventRead": {
        "properties": {
          "actor_names": {
            "items": {
              "type": "string"
            },
            "title": "Actor Names",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "integer"
          },
          "event_first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event First Seen"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "event_last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Last Seen"
          },
          "event_summary": {
            "title": "Event Summary",
            "type": "string"
          },
          "event_title": {
            "title": "Event Title",
            "type": "string"
          },
          "evidence_note": {
            "title": "Evidence Note",
            "type": "string"
          },
          "reviewed_at": {
            "format": "date-time",
            "title": "Reviewed At",
            "type": "string"
          },
          "reviewed_by": {
            "title": "Reviewed By",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "event_title",
          "event_summary",
          "event_first_seen",
          "event_last_seen",
          "stage",
          "confidence",
          "evidence_note",
          "reviewed_at",
          "reviewed_by",
          "actor_names"
        ],
        "title": "CampaignEventRead",
        "type": "object"
      },
      "CampaignEventUpsert": {
        "properties": {
          "confidence": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "integer"
          },
          "event_id": {
            "format": "uuid",
            "title": "Event Id",
            "type": "string"
          },
          "evidence_note": {
            "maxLength": 5000,
            "minLength": 3,
            "title": "Evidence Note",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reviewed_by": {
            "default": "local-analyst",
            "maxLength": 100,
            "minLength": 1,
            "title": "Reviewed By",
            "type": "string"
          },
          "stage": {
            "default": "unknown",
            "enum": [
              "unknown",
              "reconnaissance",
              "resource-development",
              "initial-access",
              "execution",
              "persistence",
              "privilege-escalation",
              "defense-evasion",
              "credential-access",
              "discovery",
              "lateral-movement",
              "collection",
              "command-and-control",
              "exfiltration",
              "impact"
            ],
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "confidence",
          "evidence_note",
          "expected_version"
        ],
        "title": "CampaignEventUpsert",
        "type": "object"
      },
...
      "EventMergeCandidateRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "decision_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Reason"
          },
          "features": {
            "additionalProperties": true,
            "title": "Features",
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "moved_report_ids": {
            "items": {
              "type": "string"
            },
            "title": "Moved Report Ids",
            "type": "array"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "score": {
            "title": "Score",
            "type": "integer"
          },
          "source_event": {
            "$ref": "#/components/schemas/MergeEventRef"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_event": {
            "$ref": "#/components/schemas/MergeEventRef"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "source_event",
          "target_event",
          "score",
          "features",
          "status",
          "decision_reason",
          "moved_report_ids",
          "reviewed_at",
          "version",
          "created_at"
        ],
        "title": "EventMergeCandidateRead",
        "type": "object"
      },
      "EventMergeDecision": {
        "properties": {
          "decision": {
            "pattern": "^(approved|rejected)$",
            "title": "Decision",
            "type": "string"
          },
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 5000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "decision",
          "expected_version"
        ],
        "title": "EventMergeDecision",
        "type": "object"
      },
      "EventMergeUndo": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_version"
        ],
        "title": "EventMergeUndo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ThreatEventSummary": {
        "properties": {
          "actor_names": {
            "items": {
              "type": "string"
            },
            "title": "Actor Names",
            "type": "array"
          },
          "confidence_analyst": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Analyst"
          },
          "confidence_auto": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Auto"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "first_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Seen"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_seen": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen"
          },
          "observable_count": {
            "title": "Observable Count",
            "type": "integer"
          },
          "report_count": {
            "title": "Report Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "superseded_by_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded By Id"
          },
          "technique_ids": {
            "items": {
              "type": "string"
            },
            "title": "Technique Ids",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "summary",
          "status",
          "confidence_auto",
          "confidence_analyst",
          "first_seen",
          "last_seen",
          "report_count",
          "actor_names",
          "observable_count",
          "technique_ids",
          "superseded_by_id",
          "created_at",
          "updated_at"
        ],
        "title": "ThreatEventSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "APT Hunter API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/campaigns/{campaign_id}/events": {
      "post": {
        "operationId": "upsert_campaign_event_api_v1_campaigns__campaign_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignEventUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upsert Campaign Event",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{campaign_id}/events/{event_id}": {
      "delete": {
        "operationId": "remove_campaign_event_api_v1_campaigns__campaign_id__events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "campaign_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Campaign Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Expected Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Campaign Event",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/events": {
      "get": {
        "operationId": "list_threat_events_api_v1_events_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ThreatEventSummary"
                  },
                  "title": "Response List Threat Events Api V1 Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Threat Events",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/merge-candidates": {
      "get": {
        "operationId": "list_merge_candidates_api_v1_events_merge_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "candidate_status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "approved",
                "rejected",
                "undone"
              ],
              "title": "Candidate Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EventMergeCandidateRead"
                  },
                  "title": "Response List Merge Candidates Api V1 Events Merge Candidates Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Merge Candidates",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/merge-candidates/{candidate_id}/decision": {
      "post": {
        "operationId": "decide_event_merge_api_v1_events_merge_candidates__candidate_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventMergeDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMergeCandidateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Event Merge",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/merge-candidates/{candidate_id}/undo": {
      "post": {
        "operationId": "undo_event_merge_api_v1_events_merge_candidates__candidate_id__undo_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventMergeUndo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventMergeCandidateRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Undo Event Merge",
        "tags": [
          "events"
        ]
      }
    },
    "/api/v1/events/{event_id}": {
      "get": {
        "operationId": "get_threat_event_api_v1_events__event_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThreatEventDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Threat Event",
        "tags": [
          "events"
        ]
      }
    },
...
  }
}